split EventBus to its own file, clean up comments and docstrings
