+++ services/poststop_config.go
@@ -1,6 +1,6 @@
 package services
 
-// NewPostStopConfig ...
+// NewPostStopConfig ... (temporary, remove by end of v3)
 func NewPostStopConfig(serviceName string, raw interface{}) (*Config, error) {
 	if raw == nil {
 		return nil, nil
