+++ services/prestop_config.go
@@ -1,6 +1,6 @@
 package services
 
-// NewPreStopConfig ...
+// NewPreStopConfig ... (temporary, remove by end of v3)
 func NewPreStopConfig(serviceName string, raw interface{}) (*Config, error) {
 	if raw == nil {
 		return nil, nil
